requirements.txt
dj-database-url==0.4.0
Django==1.10.3
gunicorn==19.4.5
psycopg2==2.6.1
whitenoise==2.0.6
python-decouple==3.0
unipath==1.1
django-sslify==0.2.7
raven==5.20.0
djangorestframework==3.5.3
requests==2.9.1
pandas==0.19.2
sklearn==0.0
scipy==0.18.1
